swipesavvy-ai-agents/requirements.txt
# Together.AI SDK
together>=1.2.0

# OpenAI (for embeddings)
openai>=1.0.0

# Embedding models
sentence-transformers>=2.2.0  # Local embeddings
tiktoken>=0.5.0  # Token counting

# Database
psycopg2-binary>=2.9.9
sqlalchemy>=2.0.0
redis>=5.0.0

# Vector databases
pinecone-client>=3.0.0  # If using Pinecone
pgvector>=0.2.0  # If using pgvector

# Web framework
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.5.0

# PII detection and security
presidio-analyzer>=2.2.0
presidio-anonymizer>=2.2.0
spacy>=3.7.0

# Utilities
python-dotenv>=1.0.0
requests>=2.31.0
httpx>=0.25.0

# Monitoring and logging
sentry-sdk>=1.38.0
python-json-logger>=2.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.12.0

# Code quality
black>=23.0.0
flake8>=6.1.0
mypy>=1.7.0
isort>=5.12.0

# Evaluation
scikit-learn>=1.3.0
pandas>=2.1.0
numpy>=1.26.0

# JWT and auth
pyjwt>=2.8.0
cryptography>=41.0.0
passlib>=1.7.4
bcrypt>=4.1.0

# Rate limiting
slowapi>=0.1.9

# Background tasks
celery>=5.3.0

# NLP models (download separately)
# python -m spacy download en_core_web_lg
